+++ config/view_manager.config.php
@@ -1,9 +1,0 @@
-<?php
-return [
-    'template_map' => [
-        'gdpro_mail/mail/example' => __DIR__ . '/../view/mail/example.phtml',
-    ],
-    'template_path_stack' => [
-        __DIR__ . '/../view'
-    ]
-];