+++ src/Factory/MessageRendererFactory.php
@@ -1,16 +1,20 @@
 <?php
-namespace GdproMailer\Factory;
+namespace Gdpro\Mailer\Factory;
 
 use Zend\ServiceManager\FactoryInterface;
 use Zend\ServiceManager\ServiceLocatorInterface;
 
+/**
+ * Class MessageRendererFactory
+ * @package Gdpro\Mailer\Factory
+ */
 class MessageRendererFactory implements FactoryInterface
 {
     public function createService(ServiceLocatorInterface $services)
     {
         $config = $services->get('config');
 
-        return new \GdproMailer\MessageRenderer(
+        return new \Gdpro\Mailer\MessageRenderer(
             $config['gdpro_mailer']['templates'],
             $services->get('viewRenderer')
         );